Move fetchers to folders
